requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==5.0
django-mptt==0.16.0
django-mptt-admin==2.6.2
executing==2.0.1
pillow==10.3.0
pure-eval==0.2.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3